Update chat.rs
As with the slack-rs crate fix the thing is being used backwards. quite common when everything is strings
I notice the tests don't really do anything that useful such as verify the input. the test only fails if the signature changes or the client is never used.
would it be acceptable to put debug asserts or at least debug logging if APIs are used obviously wrongly such as in this case? the actual thing is correct, it's just so easy to use backwards by accident.